+++ src/action-pane.ts
@@ -79,12 +79,8 @@ export class ActionPane extends LitElement {
         return html`<h2>${content}</h2>`;
       case 3:
         return html`<h3>${content}</h3>`;
-      case 4:
-        return html`<h4>${content}</h4>`;
-      case 5:
-        return html`<h5>${content}</h5>`;
       default:
-        return html`<h6>${content}</h6>`;
+        return html`<h4>${content}</h4>`;
     }
   }
 
@@ -150,9 +146,7 @@ export class ActionPane extends LitElement {
     h1,
     h2,
     h3,
-    h4,
-    h5,
-    h6 {
+    h4 {
       color: var(--mdc-theme-on-surface);
       font-family: 'Roboto', sans-serif;
       font-weight: 300;