+++ src/editors/substation/voltage-level-editor.ts
@@ -88,7 +88,7 @@ export class VoltageLevelEditor extends LitElement {
 
   /** Opens a [[`WizardDialog`]] for editing `LNode` connections. */
   openLNodeWizard(): void {
-    const wizard = wizards['LNode'].edit(this.element);
+    const wizard = wizards['LNode'].create(this.element);
     if (wizard) this.dispatchEvent(newWizardEvent(wizard));
   }
 
@@ -113,6 +113,20 @@ export class VoltageLevelEditor extends LitElement {
 
   firstUpdated(): void {
     this.addMenu.anchor = <HTMLElement>this.addButton;
+  }
+
+  private renderLNodes(): TemplateResult {
+    if (!this.showfunctions) return html``;
+
+    const lNodes = getChildElementsByTagName(this.element, 'LNode');
+
+    return lNodes.length
+      ? html`<div class="container lnode">
+          ${lNodes.map(
+            lNode => html`<l-node-editor .element=${lNode}></l-node-editor>`
+          )}
+        </div>`
+      : html``;
   }
 
   renderFunctions(): TemplateResult {
@@ -218,7 +232,7 @@ export class VoltageLevelEditor extends LitElement {
           >${this.renderAddButtons()}</mwc-menu
         >
       </abbr>
-      ${this.renderIedContainer()}${this.renderFunctions()}
+      ${this.renderIedContainer()}${this.renderLNodes()}${this.renderFunctions()}
       ${this.renderPowerTransformerContainer()}
       <div id="bayContainer">
         ${Array.from(this.element?.querySelectorAll(selectors.Bay) ?? []).map(