+++ test/unit/editors/ied/__snapshots__/da-wizard.test.snap.js
@@ -3,7 +3,7 @@ export const snapshots = {};
 
 snapshots["with no ancestors looks like the latest snapshot"] = 
 `<mwc-dialog
-  defaultaction="close"
+  defaultaction="next"
   heading="[iededitor.wizard.daTitle]"
   open=""
   style="--mdc-dialog-min-width:calc(100% + 0px)"
@@ -139,7 +139,7 @@ snapshots["with no ancestors looks like the latest snapshot"] =
 
 snapshots["with a DA element looks like the latest snapshot"] = 
 `<mwc-dialog
-  defaultaction="close"
+  defaultaction="next"
   heading="[iededitor.wizard.daTitle]"
   open=""
   style="--mdc-dialog-min-width:calc(100% + 0px)"
@@ -275,7 +275,7 @@ snapshots["with a DA element looks like the latest snapshot"] =
 
 snapshots["with a BDA element looks like the latest snapshot"] = 
 `<mwc-dialog
-  defaultaction="close"
+  defaultaction="next"
   heading="[iededitor.wizard.daTitle]"
   open=""
   style="--mdc-dialog-min-width:calc(100% + 0px)"
@@ -411,7 +411,7 @@ snapshots["with a BDA element looks like the latest snapshot"] =
 
 snapshots["with a DA element and DAI Element looks like the latest snapshot"] = 
 `<mwc-dialog
-  defaultaction="close"
+  defaultaction="next"
   heading="[iededitor.wizard.daTitle]"
   open=""
   style="--mdc-dialog-min-width:calc(100% + 0px)"
@@ -547,7 +547,7 @@ snapshots["with a DA element and DAI Element looks like the latest snapshot"] =
 
 snapshots["with a DO element looks like the latest snapshot"] = 
 `<mwc-dialog
-  defaultaction="close"
+  defaultaction="next"
   heading="[iededitor.wizard.doTitle]"
   open=""
 >
@@ -649,7 +649,7 @@ snapshots["with a DO element looks like the latest snapshot"] =
 
 snapshots["with a DO element and DOI Element looks like the latest snapshot"] = 
 `<mwc-dialog
-  defaultaction="close"
+  defaultaction="next"
   heading="[iededitor.wizard.doTitle]"
   open=""
 >